+++ src/main/kotlin/com/autonomousapps/internal/JarAnalyzer.kt
@@ -18,7 +18,8 @@ import java.util.zip.ZipFile
  * Used by [com.autonomousapps.tasks.AnalyzeJarTask].
  */
 internal class JarAnalyzer(
-  private val configuration: Configuration,
+  private val compileClasspath: Configuration,
+  private val testCompileClasspath: Configuration,
   private val artifacts: List<Artifact>,
   private val androidLinters: Set<AndroidLinterDependency>,
   private val logger: Logger,
@@ -31,12 +32,14 @@ internal class JarAnalyzer(
   }
 
   private fun computeTransitivity() {
-    val directDependencies = configuration.directDependencies()
+    val directDependencies = compileClasspath.directDependencies()
+    val testDirectDependencies = testCompileClasspath.directDependencies()
 
     // "Artifacts" are everything used to compile the project. If there is a direct artifact with a
     // matching identifier, then that artifact is NOT transitive. Otherwise, it IS transitive.
     artifacts.forEach { artifact ->
-      artifact.isTransitive = directDependencies.none { it == artifact.dependency }
+      artifact.isTransitive = directDependencies.none { it == artifact.dependency } &&
+          testDirectDependencies.none { it == artifact.dependency }
     }
   }